feat: added router + subsetting enhancements
