+++ src/icefabric/hydrofabric/__init__.py
@@ -2,5 +2,6 @@
 
 from .origin import find_origin
 from .subset import subset_hydrofabric
+from .subset_nhf import subset_nhf
 
-__all__ = ["find_origin", "subset_hydrofabric"]
+__all__ = ["find_origin", "subset_hydrofabric", "subset_nhf"]
